Python code for model training
The code will use the data.json and config.json files in Python and will train the model with the tensorflow_embedding pipeline.

Calling the train_horoscopebot method with the necessary parameters will train the model. At the end of the execution /models wil created.
Calling the predict_intent method and passing a text, we can see how our trained model does.